Avoid dead code
The es5 version will be optimized in many engines to eliminate the 'a' and 'b' setting as dead code. It would be better to at least refer to them somehow, to prevent the DCE effects from washing out the es5 vs es6 difference. (eg in Firefox the es5 version will get DCE but the es6 version won't, which is a real difference but unlikely to be relevant for actual code.)